First stab: isolating the datetime in a new crate to use it in query-ast
