+++ quickwit/Cargo.toml
@@ -13,6 +13,7 @@ members = [
   "quickwit-control-plane",
   "quickwit-core",
   "quickwit-directories",
+  "quickwit-datetime",
   "quickwit-doc-mapper",
   "quickwit-grpc-clients",
   "quickwit-indexing",
@@ -205,6 +206,7 @@ quickwit-common = { version = "0.6.0", path = "./quickwit-common" }
 quickwit-config = { version = "0.6.0", path = "./quickwit-config" }
 quickwit-control-plane = { version = "0.6.0", path = "./quickwit-control-plane" }
 quickwit-core = { version = "0.6.0", path = "./quickwit-core" }
+quickwit-datetime = { version = "0.6.0", path = "./quickwit-datetime" }
 quickwit-directories = { version = "0.6.0", path = "./quickwit-directories" }
 quickwit-doc-mapper = { version = "0.6.0", path = "./quickwit-doc-mapper" }
 quickwit-grpc-clients = { version = "0.6.0", path = "./quickwit-grpc-clients" }
