+++ quickwit/quickwit-doc-mapper/src/default_doc_mapper/mod.rs
@@ -17,8 +17,6 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
 
-mod date_time_format;
-mod date_time_parsing;
 mod date_time_type;
 mod default_mapper;
 mod default_mapper_builder;
